[ll] core: Remove outdated shade module
